preparations for first remote host attempt

+++ engine/src/main/java/com/etherblood/luna/engine/PlayerInput.java
@@ -1,7 +1,7 @@
 package com.etherblood.luna.engine;
 
 public record PlayerInput(
-        int player,
+        long player,
         Direction direction,
         ActorAction action
 ) {
